backend/requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.3.0
pydantic[email]>=2.0.0
